# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Person API
  description: The REST-ful web service with CRUD, filter, sort anf groping operation.
  version: '1.0'
servers:
  - url: http://localhost:8081
    description: Generated server url
paths:
...
  "/api/v1/persons":
    get:
      tags:
        - person
      operationId: getAllPersons
      description: Get all person
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    put:
      tags:
        - person
      operationId: updatePerson
      description: Merge person data from dto to entity
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PersonDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    post:
      tags:
        - person
      operationId: createPerson
      description: Create Person
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PersonDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/v1/addresses":
    get:
      tags:
        - address
      operationId: getAllAddress
      description: Get all address
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    put:
      tags:
        - address
      operationId: updateAddress
      description: Merge address data from dto to entity
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddressDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    post:
      tags:
        - address
      operationId: createAddress
      description: Create address
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddressDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
  "/api/v1/persons/filter":
    post:
      tags:
        - person
      operationId: getPersonsFiltered
      description: Operation to filter and sort data on any filed (firstName, address1, countryCode, state), select multiple values for sampling
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FilterDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/v1/persons/{personsId}":
    get:
      tags:
        - person
      operationId: getPersonById
      parameters:
        - name: personsId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    delete:
      tags:
        - person
      operationId: deletePerson
      description: Delete person
      parameters:
        - name: personsId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
  "/api/v1/addresses/{addressId}":
    get:
      tags:
        - address
      operationId: getAddressById
      description: Get address by id
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
    delete:
      tags:
        - address
      operationId: deleteAddress
      description: Delete adress by id
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '405':
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '406':
          description: Not Acceptable
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '415':
          description: Unsupported Media Type
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
        '500':
          description: Internal Server Error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorMessage"
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address1:
          type: string
        address2:
          type: string
        postCode:
          type: string
        city:
          "$ref": "#/components/schemas/City"
    City:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        state:
          "$ref": "#/components/schemas/State"
    Country:
      type: object
      properties:
        id:
          type: integer
          format: int64
        countryCode:
          type: string
    FilterDataDto:
      type: object
      properties:
        sorting:
          description: enum to sort value
          type: string
          enum:
            - ASC
            - DESC
        value:
          description: list of avaliable value filed to sample data from entry
          example: [ Glasgow, Southend ]
          type: array
          items:
            type: string
        select:
          description: list of selected value from list of value for building query
          type: array
          example: [ Glasgow, Southend ]
          items:
            type: string
        filterName:
          type: string
          description: name of field to filter data on this field
          example: CITY
          enum:
            - FIRST_NAME
            - SURNAME
            - ADDRESS1
            - ADDRESS2
            - CITY
            - STATE
            - POSTCODE
            - COUNTRY_CODE
            - GENDER
            - DATE_OF_BIRTH
        filterType:
          example: SELECT
          description: filter depends on the data format.Select - string value(name, surname). Range - number format(price, date)
          type: string
          enum:
            - SELECT
            - RANGE
        groupBy:
          example: false
          type: boolean
        sort:
          type: boolean
    FilterDto:
      type: object
      properties:
        result:
          description: filtering and sorting result
          "$ref": "#/components/schemas/PageImplPerson"
        filterDataDto:
          type: array
          description: list of filter properties for each field in entry
          items:
            "$ref": "#/components/schemas/FilterDataDto"
        selectedField:
          type: string
    PageImplPerson:
      type: object
      properties:
        content:
          example: [ ]
          description: filtering and sorting result
          type: array
          items:
            "$ref": "#/components/schemas/Person"
        number:
          example: 0
          type: integer
          format: int32
        size:
          example: 10
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
    Person:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        surname:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        address:
          type: array
          items:
            "$ref": "#/components/schemas/Address"
...
    State:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country:
          "$ref": "#/components/schemas/Country"
    ErrorMessage:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        description:
          type: string
    CountryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        countryCode:
          type: string
    StateDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country:
          "$ref": "#/components/schemas/CountryDto"
    AddressDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address1:
          type: string
        address2:
          type: string
        postCode:
          type: string
        city:
          "$ref": "#/components/schemas/CityDto"
    CityDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        state:
          "$ref": "#/components/schemas/StateDto"
    PersonDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        surname:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        address:
          type: array
          items:
            "$ref": "#/components/schemas/AddressDto"